backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.12
googlesearch-python==1.3.0
greenlet==3.2.0
h11==0.14.0
httptools==0.6.4
idna==3.10
playwright==1.51.0
playwright-stealth==1.0.6
pydantic==2.11.3
pydantic_core==2.33.1
pyee==12.1.1
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
setuptools==78.1.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.1
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
